Add a rough script to look up local ip addresses.
First stab at a script to launch stund on an VM instance.
The idea is to pass this script to cloud-init to get the
service going.

This queries the instance metadata interface to determine
the local ip addresses and then (pretends to) start stund
with the first two results as arguments.

We could also just awk the output of 'ip addr'.